POC/load to pinecone /requirements.txt
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.1.1
tiktoken>=0.5.1
pinecone
PyMuPDF>=1.22.5
pypandoc>=1.11
python-frontmatter>=1.0.0
tqdm>=4.66.1
mistralai>=0.0.7
numpy>=1.24.0
pillow>=10.0.0
pytest>=7.4.0
# For handling HTTP requests for files
requests>=2.31.0
# For handling base64
pillow>=10.0.0
langchain-community